refactor: change file name interface
