+++ src/pages/FillSubmit/components/RowRadioGroupComponent/RowRadioGroupComponent.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames/bind';
 import { Radio } from '@mui/material';
 import { useAppDispatch, useAppSelector } from '../../../../redux/store';
-import RowInterface from '../../../../utils/interfaces/Row';
+import { RowInterface } from '../../../../utils/interfaces';
 import { setMultiChooseGColumn } from '../../../../redux/slice/submitform';
 
 interface Props {
