+++ src/utils/interfaces/GColumnInterface.ts
