+++ src/utils/interfaces/RowInterface.ts
