+++ src/utils/interfaces/SurveyDataInterface.ts
