+++ src/utils/interfaces/ValidationInterface.ts
