+++ src/utils/interfaces/index.ts
@@ -1,0 +1,16 @@
+export { default as AnswerInterface } from './AnswerInterface';
+export { default as GColumnInterface } from './GColumnInterface';
+export { default as LinearScaleInterface } from './LinearScale';
+export { default as OptionInterface } from './OptionInterface';
+export { default as QuestionResponseInterface } from './QuestionResponse';
+export { default as QuestionInterface } from './QuestionInterface';
+export { default as QuestionType } from './QuestionType';
+export { default as ResponseInterface } from './ResponseInterface';
+export { default as RowInterface } from './RowInterface';
+export { default as SharedSurveyData } from './SharedSurveyInterface';
+export { default as SharedUserInterface } from './SharedUserInterface';
+export { default as InfoSubmitInterface } from './InfoSubmitInterface';
+export { default as SubmitFormInterface } from './SubmitFormInterface';
+export { default as SurveyInterface } from './SurveyInterface';
+export { default as SurveyData } from './SurveyDataInterface';
+export { default as UserInterface } from './UserInterface';
